+++ src/i18n/en.ts
@@ -5,17 +5,19 @@ export const EN: LocalProperty = {
   Tab_SidebarIcons: "Sidebar Workspace Icons",
   SidebarIcons_FileName: "Enter File Name",
   Tab_FolderIcons: "Folder Icons",
+  FolderIcons_Default: "Set Default Icon for New Folders",
   FolderIcons_FileName: "Enter Folder Path",
   Tab_About: "About",
   IconID: "{num}",
   IconImg: "Enter Icon Path",
   Type_Custom: "Custom",
   Type_Lucide: "lucide",
-  RemoveIconButton: "Remove",
-  AddIconButton: "Add New Icon",
   About: 
   `
   <h5>Brief Introduction</h5>
+  <ul>
+    <li>The plugin will automatically generate the CustomIcon-AutoGen.css file in the snippets folder. Please do not modify this file as any changes will be futile (doge).</li>
+  </ul>
   <ul>
     <li>Support for custom icons: Online URLs or local relative (absolute) file paths or base64 encoding or SVG tags. For more details, see <a href="https://github.com/RavenHogWarts/obsidian-custom-sidebar-icons/blob/master/README_ZH.md" target="_blank">README</a>.</li>
     <li>Direct use of lucide source icon names is supported. For more details, see <a href="https://lucide.dev/icons/" target="_blank">lucide</a>.</li>