+++ src/i18n/types.ts
@@ -3,6 +3,7 @@ export interface LocalProperty {
     Tab_SidebarIcons: string;
     SidebarIcons_FileName: string;
     Tab_FolderIcons: string;
+    FolderIcons_Default: string;
     FolderIcons_FileName: string;
     Tab_About: string;
     About: string;
@@ -10,6 +11,4 @@ export interface LocalProperty {
     IconImg: string;
     Type_Custom: string;
     Type_Lucide: string;
-    RemoveIconButton: string;
-    AddIconButton: string;
 }