+++ src/i18n/zh.ts
@@ -5,17 +5,19 @@ export const ZH: LocalProperty = {
   Tab_SidebarIcons: "侧边工作区图标",
   SidebarIcons_FileName: "输入文件名称",
   Tab_FolderIcons: "文件夹图标",
+  FolderIcons_Default: "设置新建文件夹图标默认值",
   FolderIcons_FileName: "输入文件夹路径",
   Tab_About: "关于",
   IconID: "{num}",
   IconImg: "输入图标路径",
   Type_Custom: "自定义",
   Type_Lucide: "lucide",
-  RemoveIconButton: "移除",
-  AddIconButton: "添加新图标",
   About: 
   `
   <h5>简要说明</h5>
+  <ul>
+    <li>插件会自动在snippets文件夹中生成CustomIcon-AutoGen.css文件。请勿对此文件进行修改，因为修改了也没用（doge）</li>
+  </ul>
   <ul>
     <li>自定义图标支持: 在线URL 或 本地相对(绝对)文件路径 或 base64编码 或 svg标签。 详情查看<a href="https://github.com/RavenHogWarts/obsidian-custom-sidebar-icons/blob/master/README_ZH.md" target="_blank">README</a>。</li>
     <li>lucide源图标支持直接使用lucide图标名称。 详情查看<a href="https://lucide.dev/icons/" target="_blank">lucide</a>。</li>